+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Improvements
   ".nxs".
 - Added an options parser to the GUI starter script to set the state import
   at calling time with an argument. Use '-restore_state None | exit | saved'
+- Added lazy imports of pyFAI to improve startup time.
 
 Programmatic changes
 --------------------